+++ src/test/behavior.c2.test.ts
@@ -1118,36 +1118,17 @@ describe("Behavior Tests", () => {
 
         const params = (name: string) => querySig(project, name).parameters?.map((p) => p.name);
 
-        equal(params("functionWithADestructuredParameter"), [
-            "destructuredParam",
-        ]);
+        equal(params("singleParam"), ["params"]);
 
-        equal(params("functionWithADestructuredParameterAndExtraParameters"), [
-            "__namedParameters",
-            "extraParameter",
-        ]);
+        equal(params("extraParam"), ["params", "extraParameter"]);
 
-        equal(
-            params(
-                "functionWithADestructuredParameterAndAnExtraParamDirective",
-            ),
-            ["__namedParameters"],
+        equal(params("extraParamComment"), ["params"]);
+
+        equal(params("multiParam"), ["params", "params2", "__namedParameters"]);
+
+        logger.expectMessage(
+            'warn: The signature extraParamComment has an @param with name "fakeParameter", which was not used',
         );
-
-        const logs = [
-            'warn: The signature functionWithADestructuredParameterAndExtraParameters has an @param with name "destructuredParam", which was not used',
-            'warn: The signature functionWithADestructuredParameterAndExtraParameters has an @param with name "destructuredParam.paramZ", which was not used',
-            'warn: The signature functionWithADestructuredParameterAndExtraParameters has an @param with name "destructuredParam.paramG", which was not used',
-            'warn: The signature functionWithADestructuredParameterAndExtraParameters has an @param with name "destructuredParam.paramA", which was not used',
-            'warn: The signature functionWithADestructuredParameterAndAnExtraParamDirective has an @param with name "fakeParameter", which was not used',
-            'warn: The signature functionWithADestructuredParameterAndAnExtraParamDirective has an @param with name "destructuredParam", which was not used',
-            'warn: The signature functionWithADestructuredParameterAndAnExtraParamDirective has an @param with name "destructuredParam.paramZ", which was not used',
-            'warn: The signature functionWithADestructuredParameterAndAnExtraParamDirective has an @param with name "destructuredParam.paramG", which was not used',
-            'warn: The signature functionWithADestructuredParameterAndAnExtraParamDirective has an @param with name "destructuredParam.paramA", which was not used',
-        ];
-        for (const log of logs) {
-            logger.expectMessage(log);
-        }
         logger.expectNoOtherMessages();
     });