+++ src/test/converter2/issues/gh3026.js
@@ -1,0 +1,7 @@
+/**
+ * @param {object} params params
+ * @param {string} params.y y
+ * @returns {Promise<void>}
+ * @this {{ x: string }} this param
+ */
+export async function extendedResponse({ y }) {}